+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/ComputedStat.java
@@ -1,7 +1,7 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class ComputedStat {
-    public double statHash;
+    public long statHash;
     public int value;
     public int minimum;
     public int maximum;
