+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/Dyes.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class Dyes {
-    public double channelHash;
-    public double dyeHash;
+    public long channelHash;
+    public long dyeHash;
 }
