+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/RewardItem.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class RewardItem {
-    public double itemHash;
+    public long itemHash;
     public int value;
 }
