+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/ScaledStat.java
@@ -1,7 +1,7 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class ScaledStat {
-    public double statHash;
+    public long statHash;
     public int maximumValue;
     public boolean displayAsNumeric;
     public DisplayInterpolation[] displayInterpolation;
