+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/ActivationRequirement.java
@@ -1,5 +1,5 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 
 public class ActivationRequirement extends Requirement {
-    public double[] materialRequirementHashes;
+    public long[] materialRequirementHashes;
 }
