+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/GenderDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class GenderDefinition extends BaseDefinition {
-    public double genderHash;
+    public long genderHash;
     public int genderType;
     public String genderName;
     public String genderDescription;
