+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/OverrideDefinition.java
@@ -2,7 +2,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 
 // TODO: it does not look like Definition
 public class OverrideDefinition {
-    public double statHash;
+    public long statHash;
     public String displayName;
     public String displayDescription;
     public String displayIcon;
