+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/PlaceDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class PlaceDefinition extends BaseDefinition {
-    public double placeHash;
+    public long placeHash;
     public String placeName;
     public String placeDescription;
     public String icon;
