+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/activity_history_stats/ActivityDetails.java
@@ -1,8 +1,8 @@
 package com.soramusoka.destinyApiClient.dto_layer.activity_history_stats;
 
 public class ActivityDetails {
-    public double referenceId;
+    public long referenceId;
     public String instanceId;
     public int mode;
-    public double activityTypeHashOverride;
+    public long activityTypeHashOverride;
 }
