+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/common/StatItem.java
@@ -1,7 +1,7 @@
 package com.soramusoka.destinyApiClient.dto_layer.common;
 
 public class StatItem {
-    public double statHash;
+    public long statHash;
     public int value;
     public int maximumValue;
 }
