+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/inventory_item/InventoryItemPerk.java
@@ -2,6 +2,6 @@ package com.soramusoka.destinyApiClient.dto_layer.inventory_item;
 
 public class InventoryItemPerk {
     public String iconPath;
-    public double perkHash;
+    public long perkHash;
     public boolean isActive;
 }
